+++ pmml-evaluator/src/main/java/org/jpmml/evaluator/ProbabilityAggregator.java
@@ -43,7 +43,7 @@ class ProbabilityAggregator extends ClassificationAggregator<String> {
 		for(String category : categories){
 			Double probability = hasProbability.getProbability(category);
 
-			add(category, probability.doubleValue() * weight);
+			add(category, probability * weight);
 		}
 	}
